+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added
   - Support for Enums
+  - Parameter values display in netlist view
 
 ## 1.4.2 - 10/11/2025 - Latest Release